folder renaming for publish
